dim.incidence {incidence}	R Documentation

Access various elements of an incidence object

Description

Access various elements of an incidence object

Usage

## S3 method for class 'incidence'
dim(x)

get_interval(x, ...)

## Default S3 method:
get_interval(x, ...)

## S3 method for class 'incidence'
get_interval(x, integer = TRUE, ...)

get_n(x)

## Default S3 method:
get_n(x)

## S3 method for class 'incidence'
get_n(x)

get_timespan(x)

## Default S3 method:
get_timespan(x)

## S3 method for class 'incidence'
get_timespan(x)

Arguments

x
an incidence object.

...
Unused

integer
When TRUE (default), the interval will be converted to an integer vector if it is stored as a character in the incidence object.

Value

dim() the dimensions in the number of bins and number of groups

get_interval() if integer = TRUE: an integer vector, otherwise: the value stored in x$interval

get_n() The total number of cases stored in the object

get_timespan(): an integer denoting the timespan represented by the incidence object.

See Also

get_counts() to access the matrix of counts

get_dates() to access the dates on the right, left, and center of the interval.

group_names() to access and possibly re-name the groups

Examples

set.seed(999)
dat <- as.Date(Sys.Date()) + sample(-3:50, 100, replace = TRUE)
x <- incidence(dat, interval = "month")

# the value stored in the interval element
get_interval(x)

# the numeric value of the interval in days
get_interval(x, integer = FALSE)

# the number of observations in the object
get_n(x)

# the length of time represented
get_timespan(x)

# the number of groups
ncol(x)

# the number of bins (intervals)
nrow(x)

[Package incidence version 1.7.1 Index]