bootstrap {incidence}	R Documentation

Bootstrap incidence time series

Description

This function can be used to bootstrap incidence objects. Bootstrapping is done by sampling with replacement the original input dates. See details for more information on how this is implemented.

Usage

bootstrap(x, randomise_groups = FALSE)

Arguments

x
An incidence object.

randomise_groups
A logical indicating whether groups should be randomised as well in the resampling procedure; respective group sizes will be preserved, but this can be used to remove any group-specific temporal dynamics. If FALSE (default), data are resampled within groups.

Details

As original data are not stored in incidence objects, the bootstrapping is achieved by multinomial sampling of date bins weighted by their relative incidence.

Value

An incidence object.

Author(s)

Thibaut Jombart thibautjombart@gmail.com

See Also

incidence::find_peak to use estimate peak date using bootstrap

Examples

if (require(outbreaks) && require(ggplot2)) { withAutoprint({
  i <- incidence(fluH7N9_china_2013$date_of_onset)
  i
  plot(i)

  ## one simple bootstrap
  x <- bootstrap(i)
  x
  plot(x)

})}

[Package incidence version 1.7.1 Index]