cumulate {incidence}	R Documentation

Compute cumulative 'incidence'

Description

cumulate is an S3 generic to compute cumulative numbers, with methods for different types of objects:

Usage

cumulate(x)

## Default S3 method:
cumulate(x)

## S3 method for class 'incidence'
cumulate(x)

Arguments

x
An incidence object.

Details

default method is a wrapper for cumsum

incidence objects: computes cumulative incidence over time

projections objects: same, for projections objects, implemented in the similarly named package; see ?cumulate.projections for more information, after loading the package

Author(s)

Thibaut Jombart thibautjombart@gmail.com

See Also

The incidence() function to generate the 'incidence' objects.

Examples

dat <- as.integer(c(0,1,2,2,3,5,7))
group <- factor(c(1, 2, 3, 3, 3, 3, 1))
i <- incidence(dat, groups = group)
i
plot(i)

i_cum <- cumulate(i)
i_cum
plot(i_cum)

[Package incidence version 1.7.1 Index]