pool {incidence}	R Documentation

Pool 'incidence' across groups

Description

This function pools incidence across all groups of an incidence object. The resulting incidence() object will contains counts summed over all groups present in the input.

Usage

pool(x)

Arguments

x
An 'incidence' object.

Author(s)

Thibaut Jombart thibautjombart@gmail.com

See Also

The incidence() function to generate the 'incidence' objects.

Examples

dat <- as.integer(c(0,1,2,2,3,5,7))
group <- factor(c(1, 2, 3, 3, 3, 3, 1))
i <- incidence(dat, groups = group)
i
i$counts

## pool all groups
pool(i)
pool(i)$counts

## pool only groups 1 and 3
pool(i[,c(1,3)])
pool(i[,c(1,3)])$counts

[Package incidence version 1.7.1 Index]